pyproject.toml
[project]
name = "agentic-sdlc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "codecarbon[carbonboard]>=3.2.3",
    "fastapi>=0.133.0",
    "langchain-core>=1.2.15",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langgraph-prebuilt>=1.0.8",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.4",
]
